Create Ex12
Consider dividing a string into two halves. If the length is even, the front and back halves are the same length. If the length is odd, we'll say that the extra char goes in the front half.

e.g. 'abcde', the front half is 'abc', the back half 'de'.

Given 2 strings, a and b, return a string of the form a-front + b-front + a-back + b-back